+++ packages/workflow/bin/nwayo-cli.js
@@ -5,7 +5,7 @@
 const path = require("node:path");
 
 require(path.join("..", "cli", "index"))({
-	cliPkg: require(path.join("..", "package.json")),
+	cliPkg: require(path.join("..", "package.json")), // eslint-disable-line node/global-require, unicorn/prevent-abbreviations
 	cliPath: __dirname,
 	cliUsage: { tasks: {} },
 });