+++ packages/workflow/tasks/styles.js
@@ -16,12 +16,14 @@ const gulpsass = require("gulp-sass")(require("sass"));
 const sourcemaps = require("gulp-sourcemaps");
 const cloneDeep = require("lodash.clonedeep");
 const pluralize = require("pluralize");
+const postcssFunctions = require("postcss-functions");
 const slash = require("slash");
 const jsonToScss = require("@absolunet/json-to-scss");
 const stylelint = require("@ronilaukkarinen/gulp-stylelint");
 const env = require("../helpers/env"); // eslint-disable-line unicorn/prevent-abbreviations
 const flow = require("../helpers/flow");
 const paths = require("../helpers/paths");
+const customPostCSSFunctions = require("../helpers/postcss-functions");
 const toolbox = require("../helpers/toolbox");
 const util = require("../helpers/util");
 
@@ -118,8 +120,6 @@ module.exports = () => {
 		"styles-compile",
 
 		({ taskName }) => {
-			const sassFunctions = require(paths.config.sassFunctions); // eslint-disable-line node/global-require
-
 			const streams = [];
 
 			for (const name of Object.keys(env.bundles)) {
@@ -144,7 +144,12 @@ module.exports = () => {
 					const destination = `${bundle.output.build}/${paths.build.styles}`;
 					const source = `${util.getGeneratedBanner(name)}${list.join("\n")}\n`;
 
-					const postCssPlugins = [autoprefixer({ overrideBrowserslist: bundle.styles.options.autoprefixer })];
+					const postCssPlugins = [
+						postcssFunctions({
+							functions: customPostCSSFunctions,
+						}),
+						autoprefixer({ overrideBrowserslist: bundle.styles.options.autoprefixer }),
+					];
 					if (toMinify) {
 						postCssPlugins.push(
 							cssnano({
@@ -167,7 +172,6 @@ module.exports = () => {
 							.pipe(
 								gulpsass({
 									includePaths: [paths.directory.root],
-									functions: sassFunctions,
 									// sourcemaps  (bundle.styles.options.sourcemaps)
 								}).on("error", gulpsass.logError)
 							)